styling van verschillende paginas

+++ src/main/java/be/thomasmore/gunranch/repositories/ReservationRepository.java
@@ -3,6 +3,6 @@ package be.thomasmore.gunranch.repositories;
 import be.thomasmore.gunranch.model.Reservation;
 import org.springframework.data.repository.CrudRepository;
 
-public interface ReservationRepository extends CrudRepository<Reservation,Integer> {
+public interface ReservationRepository extends CrudRepository<Reservation,Long> {
 
 }
